+++ resources/clients/react/elwing/src/plugins/item-tracker/ItemTrackerStore.ts
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import create from "zustand";
 import {
   WorkItem,
