+++ resources/clients/react/elwing/src/plugins/item-tracker/RestService.ts
@@ -1,4 +1,5 @@
-import { mapObjIndexed, reduceWhile } from "ramda";
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 
 export type RestParams<T> = Partial<Record<keyof T, string>>;
