+++ resources/clients/react/elwing/src/plugins/item-tracker/config.json
@@ -1,0 +1,3 @@
+{
+  "BASE_URL": "http://localhost:8080/api"
+}
