+++ resources/clients/react/elwing/src/plugins/item-tracker/index.ts
@@ -1,7 +1,5 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 
 import { SideNavigationProps } from "@cloudscape-design/components";
 import { lazy } from "react";
@@ -14,8 +12,14 @@ const ItemTracker = {
     items: [
       {
         type: "link",
-        text: "Example external link",
-        href: "https://cloudscape.design/",
+        text: "Creating Spring RDS REST (Java)",
+        href: "https://github.com/awsdocs/aws-doc-sdk-examples/tree/main/javav2/usecases/Creating_Spring_RDS_%20Rest",
+        external: true,
+      },
+      {
+        type: "link",
+        text: "Creating Spring RDS REST ()",
+        href: "https://github.com/awsdocs/aws-doc-sdk-examples/tree/main/kotlin/usecases/serverless_rds",
         external: true,
       },
     ],
