+++ resources/clients/react/elwing/src/plugins/item-tracker/work-item/AddWorkItemComponent.tsx
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import {
   Button,
   Form,
